bpf: csum size_to type changed
The function bpf_csum_diff arguments `from_size` and `to_size` have type
ARG_CONST_SIZE_OR_ZERO which is `__u64`.
The commit changes the corresponding arguments type to `__u64` which allows
to overcome some eBPF verifier related issues.

Suggested-by: gentoo-root <maxim@isovalent.com>

+++ bpf/include/bpf/helpers.h
@@ -59,9 +59,12 @@ BPF_FUNC(trace_printk, const char *fmt, int fmt_size, ...);
 /* Random numbers */
 static __u32 BPF_FUNC(get_prandom_u32);
 
-/* Checksumming */
-static int BPF_FUNC_REMAP(csum_diff_external, const void *from, __u32 size_from,
-			  const void *to, __u32 size_to, __u32 seed) =
+/* Checksumming size_from and size_to parameters have __u64 type
+ * intentional which corresponds to ARG_CONST_SIZE_OR_ZERO type
+ * used in the eBPF verifier.
+ */
+static int BPF_FUNC_REMAP(csum_diff_external, const void *from, __u64 size_from,
+			  const void *to, __u64 size_to, __u32 seed) =
 	(void *)BPF_FUNC_csum_diff;
 
 /* Tail calls */
